refactor: renames <Input/> stories, component file and improves its composition
